requirements.txt
torch
pillow
torchvision
tqdm